requirements.txt
#aws-cdk.core==1.123.0
aws-cdk-lib==2.3.0
constructs>=10.0.0,<11.0.0